Cmd.ContextCmd

Description

Module with the functions used in context Mode.

Documentation

renderCtx :: Command -> [String]

ctxNExecFun :: Command -> Maybe String -> IO CtResult -> IO CtResult

Receives a command, might receive a logic, a preious context (CtResult) and returns a new one.

Creates the new comand and returns a new context with the previous one and the new command.

newCtxRes :: [String] -> String -> IO CtResult

ctxExecFun :: CmdPath -> Args -> Command -> IO CtResult -> IO CtResult

ctxSetLogic :: Name -> IO CtResult -> IO CtResult

ctxSetOption :: Option -> IO CtResult -> IO CtResult

ctxSetInfo :: Attr -> IO CtResult -> IO CtResult

ctxDeclareType :: String -> Name -> Integer -> IO CtResult -> IO CtResult

ctxDefineType :: String -> Name -> [Name] -> Type -> IO CtResult -> IO CtResult

ctxDeclareFun :: String -> Name -> [Type] -> Type -> IO CtResult -> IO CtResult

ctxDefineFun :: String -> Name -> [Binder] -> Type -> Expr -> IO CtResult -> IO CtResult

ctxPush :: Integer -> IO CtResult -> IO CtResult

ctxPop :: Integer -> IO CtResult -> IO CtResult

ctxAssert :: Expr -> IO CtResult -> IO CtResult

ctxCheckSat :: CmdPath -> Args -> IO CtResult -> IO CtResult

ctxGetAssertions :: CmdPath -> Args -> IO CtResult -> IO CtResult

ctxGetValue :: CmdPath -> Args -> [Expr] -> IO CtResult -> IO CtResult

ctxGetProof :: CmdPath -> Args -> IO CtResult -> IO CtResult

ctxGetUnsatCore :: CmdPath -> Args -> IO CtResult -> IO CtResult

ctxGetInfo :: CmdPath -> Args -> InfoFlag -> IO CtResult -> IO CtResult

ctxGetOption :: CmdPath -> Args -> Name -> IO CtResult -> IO CtResult

ctxExit :: CmdPath -> Args -> IO CtResult -> IO CtResult