Cmd.ProcCom.Process

Description

The following module contains method that facilitate the comunication with the SMTSolvers.

Documentation

beginProcess :: CmdPath -> Args -> IO Process

Creates a Process ready to be executed.

send :: Process -> String -> IO String

Sends the desired input to the process std_in and then reads from std out. Working smt with this method: - z3 - mathSat - cvc4

endProcess :: Process -> IO ExitCode

Sends the signal to terminate to the running process.

sendContext :: CmdPath -> Args -> Context -> IO String

It's the same function as readProcess. http://hackage.haskell.org/package/process-1.1.0.1/docs/System-Process.html

sendScript :: CmdPath -> Args -> FilePath -> IO String

Calls readProcess and pass as arguments the arguments given plus the name of the file with the context. An empty String is passed to the std_in.

It's the same function as readProcess. readProcess: http://hackage.haskell.org/package/process-1.1.0.1/docs/System-Process.html

type Process = (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)

Type returned by CreateProcess

type CmdPath = String

Path to the process

type Args = [String]

Argumants passed to process