Z3

Description

Module wich has the standard configuration for all Z3 Modes and provides the initilizing function.

Documentation

startZ3 :: Mode -> String -> Maybe SolverConfig -> Maybe FilePath -> IO Solver

Function that initialyzes a Z3 Solver. It Receives a Mode, an SMT Logic, it can receive a diferent configuration for the solver and an anternative path to create the script in Script Mode.

In Context and Online Mode if a FilePath is passed then it's ignored.